Support FreeBSD and simplify Windows-checking logic
From what I see, OsHelper::getOS is only used in midelwares.php to
determine whether or not a OS is Windows and therefore permissions
should be checked, so while at it, I add a static method to check
precisely for that.

Without this patch or something similar, the API endpoints fail on
FreeBSD for no good reason.