Update SugarFolders.php
These changes are necessary to fix multiple problems with group folders display for users. Fix isgroup vs is_group error.  Test if user has access to the group folder, ensure admin always sees the group folders in the arrays, and to remove child folders from the array as well as check for and manage duplicates so folders don't appear twice.

This change is for 7.13+