REL #250 Remove request bootstrap.min.css
The file boostrap.min.css it's missing under the theme css folder.

The file under the css folder of suite8 theme there are the boostrap
files uncompressed, and they are imported with the style.scss file of
Dawn or Noon styles, so the line of layout_utils.php it's need it.