Fix #347 - date component default value
On init of the date component, the field should default to null if empty. This is done by the condition checking function isVoid and isEmptyString. The last update to this file forgot to add an else to the next set value, that would override the previous setValue to null. This would result in an error triggered by the validation on save event. Because empty string is an invalid date. This fix will ensure the date is null or valid string date.